# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type PdfInputPageSize = typeof PdfInputPageSize[keyof typeof PdfInputPageSize];


export const PdfInputPageSize = {
  A4: 'A4',
  Letter: 'Letter',
  Legal: 'Legal',
} as const;

export type PdfInputOrientation = typeof PdfInputOrientation[keyof typeof PdfInputOrientation];


export const PdfInputOrientation = {
  portrait: 'portrait',
  landscape: 'landscape',
} as const;

export interface PdfInput {
  /**
     * @minLength 1
     * @maxLength 2048
     */
  url: string;
  pageSize: PdfInputPageSize;
  orientation: PdfInputOrientation;
  hideNotices?: boolean;
  expandedCapture?: boolean;
  requestId?: string;
}

export type PdfProgressStage = typeof PdfProgressStage[keyof typeof PdfProgressStage];


export const PdfProgressStage = {
  starting: 'starting',
  loading: 'loading',
  removing: 'removing',
  expanding: 'expanding',
  carousels: 'carousels',
  rendering: 'rendering',
  complete: 'complete',
  error: 'error',
} as const;

export interface PdfProgress {
  stage: PdfProgressStage;
  message: string;
}

export interface ApiError {
  error: string;
  code?: string;
}

export type PdfRendererHealthStatus = typeof PdfRendererHealthStatus[keyof typeof PdfRendererHealthStatus];


export const PdfRendererHealthStatus = {
  ok: 'ok',
} as const;

export type PdfRendererHealthRenderer = typeof PdfRendererHealthRenderer[keyof typeof PdfRendererHealthRenderer];


export const PdfRendererHealthRenderer = {
  ready: 'ready',
} as const;

export interface PdfRendererHealth {
  status: PdfRendererHealthStatus;
  renderer: PdfRendererHealthRenderer;
}

export interface HealthStatus {
  status: string;
}