# lib/api-zod/src/generated/types/pdfInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PdfInputOrientation } from './pdfInputOrientation';
import type { PdfInputPageSize } from './pdfInputPageSize';

export interface PdfInput {
  /**
     * @minLength 1
     * @maxLength 2048
     */
  url: string;
  pageSize: PdfInputPageSize;
  orientation: PdfInputOrientation;
  hideNotices?: boolean;
  expandedCapture?: boolean;
  requestId?: string;
}