# lib/api-zod/src/generated/types/pdfProgress.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PdfProgressStage } from './pdfProgressStage';

export interface PdfProgress {
  stage: PdfProgressStage;
  message: string;
}