# lib/api-zod/src/generated/types/pdfRendererHealth.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PdfRendererHealthRenderer } from './pdfRendererHealthRenderer';
import type { PdfRendererHealthStatus } from './pdfRendererHealthStatus';

export interface PdfRendererHealth {
  status: PdfRendererHealthStatus;
  renderer: PdfRendererHealthRenderer;
}